+++ pmd-kotlin/src/main/java/net/sourceforge/pmd/lang/kotlin/ast/AttributeView.java
@@ -77,7 +77,11 @@ public abstract class AttributeView<T extends KotlinNode> implements Node {
         return node.getChild(index);
     }
 
+    /**
+     * @deprecated Since 7.25.0. Don't use getImage()! See #4787.
+     */
     @Override
+    @Deprecated
     public boolean hasImageEqualTo(String image) {
         return node.hasImageEqualTo(image);
     }